refactor: group core and helper files and trim app logging
Introduces lib/core (bloc, sensing, CAWS backend, resource managers) and
lib/helpers (config, logging, utils), retiring lib/blocs. View model files
are renamed to *_view_model.dart and each documents, in the same shape,
which page it backs, the state it holds, and its responsibility.

Drops logAppState and its state snapshot, which dumped seventeen lines on
every router redirect, along with the temporary timing logs.
