+++ lib/core/app_bloc.dart
@@ -71,27 +71,15 @@ class AppBloc extends ChangeNotifier {
       'DeploymentMode: ${AppConfig.deploymentMode.name}, '
       'DebugLevel: ${AppConfig.debugLevel.name}',
     );
-
-    // The coordinator is the sole router-notifier - forward service changes.
-    // Once consent is given, the study can be configured and started.
-    consent.addListener(_onConsentChanged);
-  }
-
-  void _onConsentChanged() {
-    logAppState('AppBloc._onConsentChanged() - consent.isAccepted=${consent.isAccepted}');
-    notifyListeners();
-    if (consent.isAccepted == true) unawaited(tryConfigureStudy());
   }
 
   /// Run [configureStudy], surfacing a failure to the user instead of
   /// throwing. Used by the setup flow and retry affordances, where no
   /// caller can handle the error.
   Future<void> tryConfigureStudy() async {
-    logAppState('AppBloc.tryConfigureStudy() START');
     try {
       await configureStudy();
     } catch (error) {
-      logAppState('AppBloc.tryConfigureStudy() FAILED - $error');
       final context = scaffoldKey.currentContext;
       final locale = context != null ? RPLocalizations.of(context) : null;
       scaffoldKey.currentState?.showSnackBar(
@@ -103,7 +91,6 @@ class AppBloc extends ChangeNotifier {
   /// Initialize this BLOC. Called before being used for anything.
   Future<void> initialize() async {
     if (isInitialized) return;
-    logAppState('AppBloc.initialize() START');
 
     Settings().debugLevel = AppConfig.debugLevel;
     await Settings().init();
@@ -124,15 +111,6 @@ class AppBloc extends ChangeNotifier {
     _state = AppState.initialized;
     notifyListeners();
     debug('$runtimeType initialized - deployment mode: ${AppConfig.deploymentMode.name}');
-
-    logAppState('AppBloc.initialize() DONE');
-
-    // For a returning user, check consent - via [_onConsentChanged] this
-    // routes to the consent page or configures and starts the study.
-    if (study.hasStudy) {
-      logApp('AppBloc.initialize() - returning user has a persisted study, refreshing consent status');
-      unawaited(consent.refreshStatus(study.study));
-    }
   }
 
   /// Set the active study in the app based on an [invitation].
@@ -140,11 +118,6 @@ class AppBloc extends ChangeNotifier {
   /// The study translations are re-loaded by the app, which listens for the
   /// study change.
   void setStudyInvitation(ActiveParticipationInvitation invitation) {
-    logApp(
-      'AppBloc.setStudyInvitation() - accepting invitation: '
-      'deploymentId=${invitation.studyDeploymentId}, participantId=${invitation.participation.participantId}',
-    );
-
     // create and save the participant info based on this invitation
     LocalSettings().participant = Participant.fromParticipationInvitation(invitation);
 
@@ -158,11 +131,6 @@ class AppBloc extends ChangeNotifier {
     notifyListeners();
 
     info('Invitation received - study: ${study.study}');
-    logAppState('AppBloc.setStudyInvitation() DONE - study set, refreshing consent status');
-
-    // Routes to the consent page - or, if this participant has already
-    // consented (e.g. on another phone), configures and starts the study.
-    unawaited(consent.refreshStatus(study.study));
   }
 
   /// Configure the study deployment and start sensing.
@@ -179,10 +147,8 @@ class AppBloc extends ChangeNotifier {
   Future<void> configureStudy() async {
     // early out if already configuring or configured
     if (_state == AppState.configuring || isConfigured) {
-      logApp('AppBloc.configureStudy() - skipped, already ${_state.name}');
       return;
     }
-    logAppState('AppBloc.configureStudy() START');
 
     final previousState = _state;
     _state = AppState.configuring;
@@ -192,7 +158,6 @@ class AppBloc extends ChangeNotifier {
     } catch (error) {
       _state = previousState;
       warning('$runtimeType - Study configuration failed - $error');
-      logAppState('AppBloc.configureStudy() FAILED - study.configure() threw - $error');
       rethrow;
     }
 
@@ -206,7 +171,6 @@ class AppBloc extends ChangeNotifier {
 
     _state = AppState.configured;
     notifyListeners();
-    logAppState('AppBloc.configureStudy() DONE - now starting sensing');
 
     await study.start();
   }
@@ -241,7 +205,6 @@ class AppBloc extends ChangeNotifier {
     messages.stop();
     await _userTaskNotificationSubscription?.cancel();
     _userTaskNotificationSubscription = null;
-    consent.reset();
 
     // stop sensing and remove all deployment info
     await study.remove();
