+++ lib/core/resource_manager_factory.dart
