+++ lib/core/sensing.dart
