+++ lib/services/auth_service.dart
@@ -51,34 +51,7 @@ class AuthService {
             final byName = a.invitation.name.compareTo(b.invitation.name);
             return byName != 0 ? byName : a.studyDeploymentId.compareTo(b.studyDeploymentId);
           });
-    _logInvitations(all, _invitations);
     return _invitations;
-  }
-
-  /// Dump the identity of every invitation CAWS returned, so entries that look
-  /// like duplicates in the UI (same study name) can be told apart by their
-  /// deployment / participant / role ids. Filter device logs with
-  /// `flutter logs | grep INVITATIONS`.
-  void _logInvitations(List<ActiveParticipationInvitation> all, List<ActiveParticipationInvitation> kept) {
-    logApp('INVITATIONS - backend returned ${all.length}, kept ${kept.length} after the smartphone filter');
-    for (final (index, invitation) in all.indexed) {
-      logApp(
-        'INVITATIONS - [$index] ${kept.contains(invitation) ? 'KEPT   ' : 'DROPPED'} '
-        'name="${invitation.invitation.name}" '
-        'studyId=${invitation.studyId} '
-        'deploymentId=${invitation.studyDeploymentId} '
-        'participantId=${invitation.participantId} '
-        'roles=${invitation.participation.assignedRoles.roleNames} '
-        'devices=${invitation.assignedDevices?.map((d) => '${d.device.roleName}(${d.device.runtimeType})').toList()}',
-      );
-    }
-    // Same deployment twice means CAWS sent two participations of one deployment;
-    // distinct deployments with the same name means the study was deployed twice.
-    final deployments = all.map((invitation) => invitation.studyDeploymentId).toList();
-    logApp(
-      'INVITATIONS - ${deployments.toSet().length} distinct deployment(s) in ${deployments.length} invitation(s), '
-      '${all.map((invitation) => invitation.participantId).toSet().length} distinct participant id(s)',
-    );
   }
 
   /// Authenticate using a web view.
