+++ lib/view_models/participant_data_view_model.dart
@@ -1,5 +1,10 @@
 part of carp_study_app;
 
+/// View model for [ParticipantDataPage].
+///
+/// State: the participant data the study expects the user to fill in.
+///
+/// Writes the filled-in data back to the deployment.
 class ParticipantDataPageViewModel extends ViewModel {
   Set<ExpectedParticipantData?> get expectedData => bloc.study.expectedParticipantData;
