+++ lib/view_models/study_view_model.dart
@@ -1,7 +1,11 @@
 part of carp_study_app;
 
-/// The view model for the [StudyPage]. Mainly holds the list of messages like
-/// news articles to be shown as part of the study.
+/// View model for [StudyPage] and [StudyDetailsPage].
+///
+/// State: the study description and responsible party, the deployment status,
+/// and the messages (announcements, news) published for the study.
+///
+/// Also handles pull-to-refresh: re-fetch messages and re-try the deployment.
 class StudyPageViewModel extends ViewModel {
   StudyPageViewModel({
     StudyService? studyService,
