+++ lib/view_models/task_list_view_model.dart
@@ -1,6 +1,11 @@
 part of carp_study_app;
 
-/// A view model for the [TaskListPage].
+/// View model for [TaskListPage].
+///
+/// State: the user tasks to show, whether the participant-data card is due, and
+/// the task that just auto-completed.
+///
+/// Starts tasks and auto-completes background ones once they have run.
 class TaskListPageViewModel extends ViewModel {
   TaskListPageViewModel({StudyService? studyService}) : _studyService = studyService;
